Removed misleading doc.
I removed this line from the doc for mentio-search:

    The scope property referenced by <code>items</code> can be a promise.

In fact, it cannot be a promise. If you try setting it to a promise, the
template will try to iterate over it. Instead, the mentio-search function should
just set the items property as soon as it has it.

The only promise-related code in mentio is for mentio-select, not mentio-search.